requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
pgvector
nltk
pymupdf
python-docx
torch
transformers
sentence_transformers
celery>=5.0,<6.0
django-celery-results
djangorestframework